requirements.txt
streamlit>=1.38.0
google-genai>=1.0.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=7.0.0
markdownify>=0.14.1
groq>=0.25.0
